+++ flexmeasures/data/schemas/tests/test_sensor.py
@@ -20,6 +20,11 @@ def serialize_variable_quantity(value):
     return VariableQuantityDumpSchema().dump({"value": value})["value"]
 
 
+def assert_quantity_equals(quantity, expected_quantity):
+    assert str(quantity.units) == str(expected_quantity.units)
+    assert quantity.magnitude == pytest.approx(expected_quantity.magnitude)
+
+
 @pytest.mark.parametrize(
     "src_quantity, dst_unit, fails, exp_dst_quantity",
     [
@@ -93,11 +98,11 @@ def test_quantity_or_sensor_deserialize(
     try:
         dst_quantity = schema.deserialize(src_quantity)
         if isinstance(src_quantity, (ur.Quantity, int, float)):
-            assert dst_quantity == ur.Quantity(exp_dst_quantity)
-            assert str(dst_quantity) == exp_dst_quantity
+            assert_quantity_equals(dst_quantity, ur.Quantity(exp_dst_quantity))
         elif isinstance(src_quantity, list):
-            assert dst_quantity[0]["value"] == ur.Quantity(exp_dst_quantity)
-            assert str(dst_quantity[0]["value"]) == exp_dst_quantity
+            assert_quantity_equals(
+                dst_quantity[0]["value"], ur.Quantity(exp_dst_quantity)
+            )
         assert not fails
     except ValidationError as e:
         assert fails, e